Better auto colors
When changing themes in terminal vim using autocolors this plugin
currently defaults to using `darkgrey` and `black` (for dark
backgrounds) or `lightgrey` and `white` (for light backgrouns) for the
indent color. These colors often clash with otherwise harmonious themes.
In complete themes the `Normal` and `CustomCursor` background colors are
set which provide theme-appropriate indent colors.